% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddColumn2Data-nm2eV.R
\name{AddColumnEnergy}
\alias{AddColumnEnergy}
\title{AddColumnEnergy}
\usage{
AddColumnEnergy(path.to.folder)
}
\arguments{
\item{path.to.folder}{full path to folder containing data files}
}
\value{
Writes to file.
}
\description{
A poorly documented function.
To be honest I don't remember what it's for. Not exported.
}
