% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{amin2degrees}
\alias{amin2degrees}
\title{Convert arcminutes to degrees}
\usage{
amin2degrees(x)
}
\arguments{
\item{x}{arcminutes, numeric}
}
\value{
degrees, numeric
}
\description{
Convert arcminutes to degrees
}
\seealso{
https://en.wikipedia.org/w/index.php?title=Minute_and_second_of_arc&oldid=1159601591
}
