% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{as.degrees}
\alias{as.degrees}
\title{Convert from radians to degrees}
\usage{
as.degrees(radians)
}
\arguments{
\item{radians}{numeric}
}
\value{
degrees (numeric)
}
\description{
Convert from radians to degrees
}
