% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{as.radians}
\alias{as.radians}
\title{Convert from degrees to radians}
\usage{
as.radians(degrees)
}
\arguments{
\item{degrees}{numeric}
}
\value{
radians (numeric)
}
\description{
Convert from degrees to radians
}
