% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters-electrochemical.R
\name{as.SHE}
\alias{as.SHE}
\title{Convert from electrochemical or electronic scale to SHE}
\usage{
as.SHE(potential, scale, concentration = "saturated", temperature = 25,
  as.SHE.data = potentials.as.SHE())
}
\arguments{
\item{potential}{in the original scale, V or eV}

\item{scale}{name of the original scale}

\item{concentration}{of electrolyte in mol/L, or as the string "saturated"}

\item{temperature}{of system in degrees Celsius}

\item{as.SHE.data}{dataframe with dataset}
}
\value{
potential in SHE scale
}
\description{
Convert from electrochemical or electronic scale to SHE
}

