% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{asec2degrees}
\alias{asec2degrees}
\title{Convert arcseconds to degrees}
\usage{
asec2degrees(x)
}
\arguments{
\item{x}{arcseconds, numeric}
}
\value{
degrees, numeric
}
\description{
Convert arcseconds to degrees
}
\seealso{
https://en.wikipedia.org/w/index.php?title=Minute_and_second_of_arc&oldid=1159601591
}
