% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters-electrochemical.R
\name{AVS2SHE}
\alias{AVS2SHE}
\title{AVS -> SHE}
\usage{
AVS2SHE(avs)
}
\arguments{
\item{avs}{Potential in AVS scale}
}
\value{
potential in SHE scale (numeric)
}
\description{
Converts from absolute vacuum scale (AVS) to SHE scale
}

