% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{Celsius2Kelvin}
\alias{Celsius2Kelvin}
\title{Convert from Celsius scale to Kelvin}
\usage{
Celsius2Kelvin(Celsius)
}
\arguments{
\item{Celsius}{degrees Celsius (numeric)}
}
\value{
Kelvin (numeric)
}
\description{
Converts temperature from Celsius to Kelvin.
}

