% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters-electrochemical.R
\name{ConvertRefPot}
\alias{ConvertRefPot}
\title{Convert from one electrochemical scale to another}
\usage{
ConvertRefPot(argpotential, argrefscale, valuerefscale)
}
\arguments{
\item{argpotential}{potential (numeric)}

\item{argrefscale}{input reference scale (char string)}

\item{valuerefscale}{output reference scale (char string)}
}
\value{
potential in output reference scale (numeric)
}
\description{
Convert from one electrochemical scale to another
}
