% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{degrees2amin}
\alias{degrees2amin}
\title{Convert degrees to arcminutes}
\usage{
degrees2amin(x)
}
\arguments{
\item{x}{degrees, numeric}
}
\value{
arcminutes, numeric
}
\description{
Convert degrees to arcminutes
}
