% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{degrees2asec}
\alias{degrees2asec}
\title{Convert degrees to arcseconds}
\usage{
degrees2asec(x)
}
\arguments{
\item{x}{degrees, numeric}
}
\value{
arcseconds, numeric
}
\description{
Convert degrees to arcseconds
}
