% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{ExtractSampleIdString}
\alias{ExtractSampleIdString}
\title{Extract sampleid from a string}
\usage{
ExtractSampleIdString(string)
}
\arguments{
\item{string}{character string that contains a sampleid}
}
\value{
sampleid (character string)
}
\description{
Extract sampleid from a string by utilising the fact that samepleid's
adhere to a defined format.
}

