% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters-electrochemical.R
\name{from.SHE}
\alias{from.SHE}
\title{Convert from SHE scale to another electrochemical or physical scale}
\usage{
from.SHE(potential, scale, electrolyte = "", concentration = "saturated",
  temperature = 25, as.SHE.data = potentials.as.SHE())
}
\arguments{
\item{potential}{potential in volt}

\item{scale}{name of the target scale}

\item{electrolyte}{optional, specify electrolyte solution, e.g., "KCl(aq)". Must match one of the values in \code{\link{potentials.as.SHE}$electrolyte}}

\item{concentration}{of electrolyte in mol/L, or as the string "saturated"}

\item{temperature}{of system in degrees Celsius}

\item{as.SHE.data}{by default this parameter reads the full dataset \code{\link{potentials.as.SHE}}}
}
\value{
potential in the specified target scale
}
\description{
Convert an arbitrary number of potentials vs SHE to another electrochemical
scale (or the vacuum scale).
The available target scales are those listed by \code{\link{potentials.as.SHE}}.
}
