% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-xtable.R
\name{GenericXtableSetAttributes}
\alias{GenericXtableSetAttributes}
\title{Set the attributes for a generic xtable object}
\usage{
GenericXtableSetAttributes(xtobject, nxtnames = NULL, nxtdigits = NULL,
  nxtdisplay = NULL, nxtalign = NULL, caption.text = "nxtcaption",
  caption.label = "tab:nxtlabel")
}
\arguments{
\item{xtobject}{the xtable (table)}

\item{nxtnames}{vector of names (column names)}

\item{nxtdigits}{vector of digits (0 if column is non-numeric, numeric of desired number of digits otherwise)}

\item{nxtdisplay}{vector of display format [see formatC(format=...)]}

\item{nxtalign}{vector of LaTeX align (e.g., "l", "c", "r", "S[table-format=1.1]", ...)}

\item{caption.text}{string for the LaTeX caption text}

\item{caption.label}{string for the LaTeX reference label}
}
\value{
xtable object
}
\description{
This function helps you to set the attributes for an xtable
object. It returns an xtable object.
}
\details{
Sets names, digits, display, and align for the passed xtable object
}
\examples{
\dontrun{
xtabWithAttributes <- GenericXtableSetAttributes(xtobject)
xtabWithAttributes <- GenericXtableSetAttributes(xtobject, nxtdigits = c(0, 2, 2, 4))
}
}

