% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{int2padstr}
\alias{int2padstr}
\title{Create a string of specified width from an integer}
\usage{
int2padstr(ii, pchr, w)
}
\arguments{
\item{ii}{integer or vector of integers}

\item{pchr}{a padding character (e.g., "0")}

\item{w}{width of the return string (an integer)}
}
\value{
a string or a vector of strings
}
\description{
Converts an integer or a vector of integers to a string
of fixed length padded with a specified character (e.g., zeroes).
}
