% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{is.wholenumber}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{number}

\item{tol}{machine's double precision}
}
\value{
logical (TRUE or FALSE)
}
\description{
I am not even sure this function is useful any longer.
Kept for legacy purposes just in case some old code depends on it.
This function was copied from R's documentation (see ?is.integer).
}
