% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{Kelvin2Celsius}
\alias{Kelvin2Celsius}
\title{Convert from Kelvin to Celsius scale}
\usage{
Kelvin2Celsius(Kelvin)
}
\arguments{
\item{Kelvin}{(numeric)}
}
\value{
degrees Celsius (numeric)
}
\description{
Converts from temperature in Kelvin to degrees Celsius
}

