% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{LoadRData2Variable}
\alias{LoadRData2Variable}
\title{LoadRData2Variable}
\usage{
LoadRData2Variable(FullPathToRData)
}
\arguments{
\item{FullPathToRData}{path to rda file}
}
\value{
an R object, you will probable want to assign it to a variable
}
\description{
This function loads R-data file into a variable instead of into the workspace.
Works well when the R-data file contains only ONE variable.
Not tested for when the R-data file contains more than one variable.
}
