% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-xtable.R
\name{LongtableXtableHeader}
\alias{LongtableXtableHeader}
\title{Set xtable header in LaTeX longtable format}
\usage{
LongtableXtableHeader(xtobject, caption.text, caption.label)
}
\arguments{
\item{xtobject}{xtable object (table)}

\item{caption.text}{string for the LaTeX caption text}

\item{caption.label}{string for the LaTeX reference label}
}
\value{
character string (with LaTeX escaping)
}
\description{
This function creates a longtable header assuming
that the LaTeX document will use the booktabs package.
This function should not be used together with \code{booktabs = TRUE}
}
