% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry-tools.R
\name{molarity2mass}
\alias{molarity2mass}
\title{Calculate required mass of substance to dissolve}
\usage{
molarity2mass(formulamass, volume, molarity)
}
\arguments{
\item{formulamass}{of the substance (in grams per mole)}

\item{volume}{of the final solution (in liters)}

\item{molarity}{(in moles per liter)}
}
\value{
mass of substance (in grams)
}
\description{
You want to prepare a solution of known molarity and volume of
a particular substance.
This function calculates the required mass to weigh up.
}

