% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{numbers2swedish}
\alias{numbers2swedish}
\title{numbers2swedish}
\usage{
numbers2swedish(x)
}
\arguments{
\item{x}{number}
}
\value{
string
}
\description{
Converts a number into its corresponding words in Swedish
This is my own adaptation of numbers2words()
I think an even neater solution would be to have an argument
numbers2words(lang = "swe"), but that would require more coding
THIS FUNCTION NEEDS WORK
Cannot return proper grammar: "en miljon" and "ett tusen", or
"en miljon" and "två miljoner", and so on.
}
