% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{numbers2words}
\alias{numbers2words}
\title{numbers2words}
\usage{
numbers2words(x, billion = c("US", "UK"), and = if (billion == "US") ""
  else "and")
}
\arguments{
\item{x}{number}

\item{billion}{follow either US or UK usage rules}

\item{and}{follows the choice set in billion arg}
}
\value{
string
}
\description{
Converts a number into its corresponding words in English
THIS FUNCTION WAS PUBLISHED IN: R-News, vol 5, iss 1, May 2005, pp. 51
Original author: John Fox, Department of Sociology, McMaster University, Hamilton, Ontari
                 Canada L8S 4M4, 905-525-9140x23604
                 http://socserv.mcmaster.ca/jfox
http://cran.csiro.au/doc/Rnews/Rnews_2005-1.pdf
http://finzi.psych.upenn.edu/R/Rhelp02a/archive/46843.html
}
