% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{pascal2torr}
\alias{pascal2torr}
\title{Convert Pascal to Torr}
\usage{
pascal2torr(pascal)
}
\arguments{
\item{pascal}{numeric (note: please supply Pa, not kPa)}
}
\value{
torr, numeric
}
\description{
Convert Pascal to Torr
}
