% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters-electrochemical.R
\name{potentials.as.SHE}
\alias{potentials.as.SHE}
\title{Potentials as SHE}
\usage{
potentials.as.SHE()
}
\value{
tidy dataframe with the following columns
   \tabular{ll}{
   \code{electrode}     \tab reference electrode \cr
   \code{electrolyte}   \tab electrolyte \cr
   \code{conc.num}      \tab concentration of electrolyte, mol/L \cr
   \code{conc.string}   \tab concentration of electrolyte, as string, may also note temperature at which conc \cr
   \code{temp}          \tab temperature / degrees Celsius \cr
   \code{SHE}           \tab potential vs SHE / volt \cr
   \code{sid}           \tab set id, just for housekeeping inside this function \cr
   \code{reference}     \tab BibTeX reference \cr
   \code{dEdT}          \tab temperature coefficient / volt/kelvin \cr
   }
}
\description{
This function just outputs a tidy dataframe with potential vs SHE for
different scales, electrolytes, concentrations, and temperatures.
Using data from literature.
}

