% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{ProvideSampleId}
\alias{ProvideSampleId}
\title{Create sampleid from path to experiment datafile}
\usage{
ProvideSampleId(pathexpfile, implementation = "filename")
}
\arguments{
\item{pathexpfile}{path to an experiment datafile}

\item{implementation}{defaults to "old" behaviour, can also be set to "dirname"}
}
\value{
a sampleid (character string)
}
\description{
Returns a "unique" sample ID when supplied with a path to an experimental file.
The second arg is optional, defaults to "old" behaviour,
but can be set to "dirname" for another behaviour.
The second arg was added so as not to break older code.
}

