% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{roundup}
\alias{roundup}
\title{Round up to the nearest specified interval}
\usage{
roundup(x, nearest = 1000)
}
\arguments{
\item{x}{number}

\item{nearest}{nearest interval, e.g., 5, 10, 100, 1000, etc.}
}
\value{
a number
}
\description{
This function rounds UP to the nearest interval specified by "nearest"
http://stackoverflow.com/questions/6461209/how-to-round-up-to-the-nearest-10-or-100-or-x
}
