% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters-electrochemical.R
\name{SHE2AVS}
\alias{SHE2AVS}
\title{SHE -> AVS}
\usage{
SHE2AVS(she)
}
\arguments{
\item{she}{Potential in SHE scale}
}
\value{
potential in AVS scale (numeric)
}
\description{
Converts from SHE scale to absolute vacuum (AVS) scale
}
