% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalise first letter in each word in string}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{string of one or more words}
}
\value{
string
}
\description{
Copied verbatim from https://stackoverflow.com/a/6364905
Also see R doc of base::toupper()
}
