% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siunitx.R
\name{siunitx.uncertainty}
\alias{siunitx.uncertainty}
\title{siunitx.uncertainty}
\usage{
siunitx.uncertainty(quantity, uncertainty, digits = 6)
}
\arguments{
\item{quantity}{quantity}

\item{uncertainty}{uncertainty}

\item{digits}{number of digits}
}
\value{
a string suitable for use with siunitx num{} command
}
\description{
Writing number with exponent (such as scientific notation) with uncertainty using siunitx
}
