% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{SubfigureGenerator}
\alias{SubfigureGenerator}
\title{Generate LaTeX subfigure code}
\usage{
SubfigureGenerator(images, subcaptions, mainlabel = "fig:mainfig",
  perpage = 6, ncol = 2, landscape = FALSE)
}
\arguments{
\item{images}{vector with full paths to images (png-files or other LaTeX-compatible format)
to be put in a LaTeX subfigure environment}

\item{subcaptions}{vector with subcaptions for each subfigure}

\item{mainlabel}{string with LaTeX label for the main figure environment
should be set individually if SubfigureGenerator() is called more than once from the same document}

\item{perpage}{maximum number of images on one page, one A4 page fits six images with subcaptions}

\item{ncol}{LaTeX subfigure is setup with ncol columns}

\item{landscape}{set this to TRUE if pages are set in landscape mode}
}
\value{
a string with LaTeX code
}
\description{
Generate LaTeX subfigure code for a bunch of supplied image paths,
subcaptions, label and subfigure layout.
Supports splitting the figures over several pages or using landscape layout.
}
