% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{SubstrateHistory}
\alias{SubstrateHistory}
\title{Display the history of a substrate (sampleid)}
\usage{
SubstrateHistory(sampleid, matrix.rda)
}
\arguments{
\item{sampleid}{string}

\item{matrix.rda}{path to rdata file containing the sample matrix}
}
\value{
a dataframe
}
\description{
Display the history of a substrate (sampleid)
}
