% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-xtable.R
\name{TabularXtableHeader}
\alias{TabularXtableHeader}
\title{Set xtable header in LaTeX tabular format}
\usage{
TabularXtableHeader(xtobject, names.custom = NULL)
}
\arguments{
\item{xtobject}{xtable object (table)}

\item{names.custom}{Use \code{names.custom} to make more complicated headers, e.g., multiple-row}
}
\value{
character string (with LaTeX escaping)
}
\description{
This function should be used together with \code{booktabs = TRUE}.
}

