% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{thth2d}
\alias{thth2d}
\title{Calculate d-spacings from 2theta values}
\usage{
thth2d(thth, wavelength = 1.540562)
}
\arguments{
\item{thth}{vector with thth values in degrees}

\item{wavelength}{radiation wavelength in Angstrom}
}
\value{
d-spacings (numeric)
}
\description{
This function applies Bragg's law to calculate d-spacings from thth (n = 1)
}

