% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{torr2pascal}
\alias{torr2pascal}
\title{Convert Torr to Pascal}
\usage{
torr2pascal(torr)
}
\arguments{
\item{torr}{numeric}
}
\value{
pascal, numeric (note returns Pa, not kPa)
}
\description{
Convert Torr to Pascal
}
