% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry-tools.R
\name{VapourPressureWater}
\alias{VapourPressureWater}
\title{Vapour pressure of water}
\usage{
VapourPressureWater(temperature)
}
\arguments{
\item{temperature}{numeric vector, in degrees Celsius}
}
\value{
vapour pressure of water, in kilopascal
}
\description{
Vapour pressure of water as a function of temperature
This function returns the vapour pressure of water at the given
temperature(s) from the common::vapourwater dataset.
}
\examples{
\dontrun{
VapourPressureWater(45)
VapourPressureWater(c(20, 25, 45, 60))
}
}
