% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters.R
\name{wavenum2length}
\alias{wavenum2length}
\title{Convert wavenumber to wavelength}
\usage{
wavenum2length(wavenumber)
}
\arguments{
\item{wavenumber}{number or vector of numbers}
}
\value{
number or vector
}
\description{
Converts wavenumber (cm-1) to wavelength (nm)
Only valid for absolute wavenumbers, NOT delta wavenumbers (ranges)
http://www.powerstream.com/inverse-cm.htm
}
