% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters-electrochemical.R
\name{as.SHE}
\alias{as.SHE}
\title{Convert from electrochemical or physical scale to SHE}
\usage{
as.SHE(potential, scale, electrolyte = "", concentration = "saturated",
  temperature = 25, as.SHE.data = potentials.as.SHE())
}
\arguments{
\item{potential}{potential in volt}

\item{scale}{name of the original scale}

\item{electrolyte}{optional, specify electrolyte solution, e.g., "KCl(aq)". Must match value in \code{as.SHE.data$electrolyte}.}

\item{concentration}{of electrolyte in mol/L, or as the string "saturated"}

\item{temperature}{of system in degrees Celsius}

\item{as.SHE.data}{dataframe with dataset}
}
\value{
potential in SHE scale
}
\description{
Convert an arbitrary number of potentials against any known electrochemical
scale (or the electronic vacuum scale) to potential vs SHE.
}
