% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters-electrochemical.R
\name{ConvertRefPotEC}
\alias{ConvertRefPotEC}
\title{ConvertRefPotEC}
\usage{
ConvertRefPotEC(argpotential, argrefscale, valuerefscale)
}
\arguments{
\item{argpotential}{potential (numeric)}

\item{argrefscale}{input reference scale (character string)}

\item{valuerefscale}{output reference scale (character string)}
}
\value{
potential in output reference scale (numeric)
}
\description{
This function does the heavy lifting.
Converts from an electrochemical reference scale into another.
SHE:     standard hydrogen electrode
Ag/AgCl: silver silver-chloride electrode (3M KCl)
SCE:     saturated calomel electrode
}
