% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{LoadRData2Variable}
\alias{LoadRData2Variable}
\title{LoadRData2Variable}
\usage{
LoadRData2Variable(path, url = "")
}
\arguments{
\item{path}{local path to rda file (default)}

\item{url}{remote URL to rda file (note: requires explicitly specifying argument)}
}
\value{
an R object, you will probably want to assign it to a variable
}
\description{
This function loads R-data file into a variable instead of into the workspace.
Works well when the R-data file contains only ONE variable.
Not tested for when the R-data file contains more than one variable.
}
\examples{
\dontrun{
abscoeff <- LoadRData2Variable("/path/file.rda")
abscoeef <- LoadRData2Variable(url="http://public.solarchemist/data/file.rda")
}
}
