% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-converters-electrochemical.R
\name{RefCanonicalName}
\alias{RefCanonicalName}
\title{Get standardised name of reference electrode}
\usage{
RefCanonicalName(refname)
}
\arguments{
\item{refname}{string}
}
\value{
the canonical name or empty string
}
\description{
Given a reference electrode label, this function returns its canonical name
(as defined by this package).
This function tries to match against as many variations as possible for each
reference electrode.
}

